+++ contracts/interfaces/ITransmitManager.sol
@@ -3,12 +3,12 @@ pragma solidity 0.8.7;
 
 interface ITransmitManager {
     function checkTransmitter(
-        uint256 siblingSlug,
+        uint32 siblingSlug,
         bytes32 digest,
         bytes calldata signature
     ) external view returns (address, bool);
 
-    function payFees(uint256 dstSlug) external payable;
+    function payFees(uint32 dstSlug) external payable;
 
-    function getMinFees(uint256 dstSlug) external view returns (uint256);
+    function getMinFees(uint32 dstSlug) external view returns (uint256);
 }
